Refine java plugin doc and put all test matching logic into TestSelectionMatcher
