+++ subprojects/docs/src/docs/userguide/javaPlugin.adoc
@@ -630,7 +630,7 @@ Test filtering feature has following characteristic:
 
 * Fully qualified class name or fully qualified method name is supported, e.g. `org.gradle.SomeTest`, `org.gradle.SomeTest.someMethod`.
 * Simple class name or simple method name is supported if the pattern starts with an upper-case letter, e.g. `SomeTest`, `SomeTest.someMethod` (Since 4.7).
-* Wildcard '*' is supported for matching any characters, however, abusing wildcards might lead to performance problems - see instructions below.
+* Wildcard '*' is supported for matching any characters.
 * Command line option `--tests` is provided to conveniently extend the test filter for an individual Gradle execution. This is especially useful for the classic 'single test method execution' use case. When the command line option is used, the inclusions declared in the build script are still honored. That is, the command line filters are always applied on top of the filter definition in the build script. It is possible to supply multiple `--tests` options and tests matching any of those patterns will be included.
 * Gradle tries to filter the tests given the limitations of the test framework API. Some advanced, synthetic tests may not be fully compatible with filtering. However, the vast majority of tests and use cases should be handled neatly.
 * Test filtering supersedes the file-based test selection. The latter may be completely replaced in future. We will grow the test filtering API and add more kinds of filters.
@@ -679,18 +679,6 @@ gradle test --tests 'all.in.specific.package*'
 
 # specific method at specific package (recursively)
 gradle test --tests 'all.in.specific.package*.someSpecificMethod'
-```
-
-When a pattern starts with a wildcard, Gradle needs to load all classes to search for a potentially matched method,
-which may lead to very bad performance, especially with api:org.gradle.api.tasks.testing.Test#getForkEvery()[`forkEvery`].
-Therefore, to avoid unnecessary performance loss, please use a pattern starting with a wildcard (*) as little as possible.
-
-```
-gradle test --tests '*SomeTestClass.someSpecificFeature'
-
-gradle test --tests '*SomeSpecificTest'
-
-gradle test --tests '*SomeSpecificTestSuite'
 
 gradle test --tests '*IntegTest'
